Created Pause and Pause Option Menu UI with scripts for them
Created and completed the UI for the Pause Menu and the Options Menu within the Pause Menu in a separate scene called JH_UITest. I created the scripts for both of these UI elements and connected all elements of them. They have been tested all except quit game and the screen resolution. Created a small animation for the transition to the pause menu and created a mixer for the set up of the volume controls. For more information talk to me.