Fixed Pause and Option Menu Scripts, Resolution Issue and Imported Boss
I have tested the fix for the Pause Menu and Option Menu scripts I worked on with Aaron. I have fixed an issue with duplicates of the resolutions appearing in the Resolution Dropdown on build. I have fixed the resolutions issue for the game build. I have imported the boss model and the animations for it that I sourced.