Created Boss AI based on scripts but couldn't get it to work
I create a Boss Model AI based on the existing AI and added colliders to the model but the attack action doesn't deal damage, the boss model is hard to and colliders are weird due to the in-human nature of the model. There is also a glitch where the model is stuck in a loop of idle and wander.